+++ test/highlight/strings.groovy
@@ -11,3 +11,15 @@
 multi-line
 string'''
 // <- string
+
+"hi ${name}"
+// <- string
+//  ^ punctuation.special
+//        ^ punctuation.special
+
+"value $count"
+//     ^ variable
+
+"""triple ${expr}"""
+// <- string
+//        ^ punctuation.special